editor/libeditor/tests/test_undo_after_spellchecker_replaces_word.html

<!DOCTYPE html>
<html>
<head>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<div id="display"></div>
<textarea id="textarea">abc abx abc</textarea>
<pre id="test">
</pre>

<script class="testbody" type="application/javascript">
SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(() => {
  let textarea = document.getElementById("textarea");
  let editor = SpecialPowers.wrap(textarea).editor;

  let inlineSpellChecker = editor.getInlineSpellChecker(true);

  textarea.focus();

  function checkInputEvent(aEvent, aInputType, aData, aDescription) {
    ok(aEvent instanceof InputEvent,
       `"input" event should be dispatched with InputEvent interface ${aDescription}`);
    is(aEvent.cancelable, false,
       `"input" event should be never cancelable ${aDescription}`);
    is(aEvent.bubbles, true,
       `"input" event should always bubble ${aDescription}`);
    is(aEvent.inputType, aInputType,
       `inputType should be "${aInputType}" ${aDescription}`);
    is(aEvent.data, aData,
       `data should be ${aData} ${aDescription}`);
    is(aEvent.dataTransfer, null,
       `dataTransfer should be null ${aDescription}`);
  }

  let inputEvents = [];
  function onInput(aEvent) {
    inputEvents.push(aEvent);
  }

  SpecialPowers.Cu.import(
    "resource://testing-common/AsyncSpellCheckTestHelper.jsm")
  .onSpellCheck(textarea, () => {
    SimpleTest.executeSoon(() => {
      textarea.addEventListener("input", onInput);

      let misspelledWord = inlineSpellChecker.getMisspelledWord(editor.rootElement.firstChild, 5);
      is(misspelledWord.startOffset, 4,
         "Misspelled word should start from 4");
      is(misspelledWord.endOffset, 7,
         "Misspelled word should end at 7");
      inputEvents = [];
      inlineSpellChecker.replaceWord(editor.rootElement.firstChild, 5, "aux");
      is(textarea.value, "abc aux abc",
         "'abx' should be replaced with 'aux'");
      is(inputEvents.length, 1,
         'Only one "input" event should be fired when replacing a word with spellchecker');
      checkInputEvent(inputEvents[0], "insertReplacementText", "aux",
                      "when replacing a word with spellchecker");

      inputEvents = [];
      synthesizeKey("z", { accelKey: true });
      is(textarea.value, "abc abx abc",
         "'abx' should be restored by undo");
      is(inputEvents.length, 1,
         'Only one "input" event should be fired when undoing the replacing word');
      checkInputEvent(inputEvents[0], "historyUndo", null,
                      "when undoing the replacing word");

      inputEvents = [];
      synthesizeKey("z", { accelKey: true, shiftKey: true });
      is(textarea.value, "abc aux abc",
         "'aux' should be restored by redo");
      is(inputEvents.length, 1,
         'Only one "input" event should be fired when redoing the replacing word');
      checkInputEvent(inputEvents[0], "historyRedo", null,
                      "when redoing the replacing word");

      textarea.removeEventListener("input", onInput);

      SimpleTest.finish();
    });
  });
});
</script>
</body>
</html>
